Create python-training2
from math import pi: This line imports the constant value of π (pi) from the math module. π is approximately 3.14159 and is used in the calculations for the area and circumference of the circle.

radius = input('What is the circle radius ?'): This line prompts the user to enter the radius of the circle. The input() function takes a string as an argument (the prompt to the user) and waits for the user to enter a value. The value entered by the user is then stored in the variable radius.

print('The area of the circle is', float(radius) ** 2 * pi): This line calculates and prints the area of the circle. The float(radius) part converts the user input (which is initially a string) to a floating-point number so that mathematical operations can be performed on it. The formula for the area of a circle is A = πr^2, where r is the radius. So, float(radius) ** 2 calculates the square of the radius, and then it's multiplied by pi to get the area. The result is printed using the print() function.

print('The circumference of the circle is', 2 + float(radius) * pi): This line calculates and prints the circumference of the circle. The formula for the circumference is C = 2πr, where r is the radius. Similar to the previous calculation, float(radius) * pi calculates 2πr, and then 2 is added to it. The result is printed using the print() function.